Fix issue preventing callback_url to be provided as an option
If a `callback_url` were provided in the following manner:

    kraken.url('http://example.com/foo.jpg', :callback_url =>
'http://example.com/kraken_results')

Then this query:

    params.keys.include?(:callback_url)

would be false, because `keys` returns an array of strings (not symbols).

Also, if that were corrected, and @callback_url is nil, then

    params[:callback_url] = @callback_url

would set the callback_url to nil, and the resulting params sent to
Kraken would not have either a `callback_url` or a `wait` option.